requirements.txt
python-dotenv
langchain_groq
langchain_community
pypdf
langchain_text_splitters
langsmith
langchain_huggingface
astrapy
fastapi
langserve
langchain_core
langchain
langchain_text_splitters
langchain_community
uvicorn
python-multipart
cassio
datasets
tiktoken
langchain-astradb
langchain_chroma
webdriver-manager